+++ src/app/(customer)/faq/page.tsx
