+++ src/components/support/index.ts
@@ -1,1 +1,2 @@
 export * from "./support-form";
+export * from "./image-upload";
