+++ src/types/constants.ts
@@ -1,0 +1,2 @@
+export const IMAGE_MAX_SIZE = 5 * 1024 * 1024;
+export const ALLOWED_IMAGE_TYPES = ["image/jpeg", "image/jpg", "image/png"];
